+++ osmosis_aws_driver/computing_plugin.py
